+++ src/sentry/testutils/asserts.py
@@ -1,5 +1,6 @@
 from django.http import StreamingHttpResponse
 
+from sentry.integrations.types import EventLifecycleOutcome
 from sentry.models.auditlogentry import AuditLogEntry
 from sentry.models.commitfilechange import CommitFileChange
 from sentry.silo.base import SiloMode
@@ -60,3 +61,35 @@ def assert_org_audit_log_does_not_exist(**kwargs):
 
 def delete_all_org_audit_logs():
     return AuditLogEntry.objects.all().delete()
+
+
+"""
+Helper functions to assert integration SLO metrics
+"""
+
+
+def assert_halt_metric(mock_record, error_msg):
+    (event_halts,) = (
+        call for call in mock_record.mock_calls if call.args[0] == EventLifecycleOutcome.HALTED
+    )
+    if isinstance(error_msg, Exception):
+        assert isinstance(event_halts.args[1], type(error_msg))
+    else:
+        assert event_halts.args[1] == error_msg
+
+
+def assert_failure_metric(mock_record, error_msg):
+    (event_failures,) = (
+        call for call in mock_record.mock_calls if call.args[0] == EventLifecycleOutcome.FAILURE
+    )
+    if isinstance(error_msg, Exception):
+        assert isinstance(event_failures.args[1], type(error_msg))
+    else:
+        assert event_failures.args[1] == error_msg
+
+
+def assert_success_metric(mock_record):
+    event_success = (
+        call for call in mock_record.mock_calls if call.args[0] == EventLifecycleOutcome.SUCCESS
+    )
+    assert event_success
