--- tests/sentry/integrations/utils/test_assert_metrics.py
@@ -1,32 +1,0 @@
-from sentry.integrations.types import EventLifecycleOutcome
-
-"""
-Helper functions to assert integration SLO metrics
-"""
-
-
-def assert_halt_metric(mock_record, error_msg):
-    (event_halts,) = (
-        call for call in mock_record.mock_calls if call.args[0] == EventLifecycleOutcome.HALTED
-    )
-    if isinstance(error_msg, Exception):
-        assert isinstance(event_halts.args[1], type(error_msg))
-    else:
-        assert event_halts.args[1] == error_msg
-
-
-def assert_failure_metric(mock_record, error_msg):
-    (event_failures,) = (
-        call for call in mock_record.mock_calls if call.args[0] == EventLifecycleOutcome.FAILURE
-    )
-    if isinstance(error_msg, Exception):
-        assert isinstance(event_failures.args[1], type(error_msg))
-    else:
-        assert event_failures.args[1] == error_msg
-
-
-def assert_success_metric(mock_record):
-    event_success = (
-        call for call in mock_record.mock_calls if call.args[0] == EventLifecycleOutcome.SUCCESS
-    )
-    assert event_success
